refactor(engine): move network tests to sibling file under line limit
